requirements.txt
pyyaml >= 3.12
pyvmomi >= 6.5
docopt >= 0.6.2
netaddr >= 0.7.19
coloredlogs >= 5.2.0